+++ lib/inventree/part.dart
@@ -14,7 +14,17 @@ class InvenTreePartCategory extends InvenTreeModel {
   String NAME = "PartCategory";
 
   @override
-  String URL = "part/category/";
+  String get URL => "part/category/";
+
+  @override
+  Map<String, dynamic> formFields() {
+
+    return {
+      "name": {},
+      "description": {},
+      "parent": {}
+    };
+  }
 
   @override
   Map<String, String> defaultListFilters() {
@@ -68,7 +78,7 @@ class InvenTreePartTestTemplate extends InvenTreeModel {
   String NAME = "PartTestTemplate";
 
   @override
-  String URL = "part/test-template/";
+  String get URL => "part/test-template/";
 
   String get key => jsondata['key'] ?? '';
 
@@ -125,7 +135,33 @@ class InvenTreePart extends InvenTreeModel {
   String NAME = "Part";
 
   @override
-  String URL = "part/";
+  String get URL => "part/";
+
+  @override
+  Map<String, dynamic> formFields() {
+    return {
+      "name": {},
+      "description": {},
+      "IPN": {},
+      "revision": {},
+      "keywords": {},
+      "link": {},
+
+      // Parent category
+      "category": {
+      },
+
+      // Checkbox fields
+      "active": {},
+      "assembly": {},
+      "component": {},
+      "purchaseable": {},
+      "salable": {},
+      "trackable": {},
+      "is_template": {},
+      "virtual": {},
+    };
+  }
 
   @override
   Map<String, String> defaultListFilters() {