+++ lib/inventree/stock.dart
@@ -18,7 +18,7 @@ class InvenTreeStockItemTestResult extends InvenTreeModel {
   String NAME = "StockItemTestResult";
 
   @override
-  String URL = "stock/test/";
+  String get URL => "stock/test/";
 
   String get key => jsondata['key'] ?? '';
 
@@ -103,10 +103,23 @@ class InvenTreeStockItem extends InvenTreeModel {
   String NAME = "StockItem";
 
   @override
-  String URL = "stock/";
+  String get URL => "stock/";
 
   @override
   String WEB_URL = "stock/item/";
+
+  @override
+  Map<String, dynamic> formFields() {
+    return {
+      "part": {},
+      "location": {},
+      "quantity": {},
+      "status": {},
+      "batch": {},
+      "packaging": {},
+      "link": {},
+    };
+  }
 
   @override
   Map<String, String> defaultGetFilters() {
@@ -544,9 +557,18 @@ class InvenTreeStockLocation extends InvenTreeModel {
   String NAME = "StockLocation";
 
   @override
-  String URL = "stock/location/";
+  String get URL => "stock/location/";
 
   String get pathstring => jsondata['pathstring'] ?? '';
+
+  @override
+  Map<String, dynamic> formFields() {
+    return {
+      "name": {},
+      "description": {},
+      "parent": {},
+    };
+  }
 
   String get parentpathstring {
     // TODO - Drive the refactor tractor through this